Add fix to fprint out error.
Turns out this was a clever malloc error caused by bad hand calcs.
Effectively we create an array of pthread_t's that is too small, then
start reaching inside looking for tid's that do not exist. This then
(somehow) effects the first string to be printed to stdout. Discovered this by binary searching
with print statements for the source, then realizing mistake with the
malloc. Fixing this solved the problem. God I hate this language.